pyproject.toml
[tool.poetry]
name = "airflow-series"
version = "0.1.0"
description = ""
authors = ["Levy Nunes <levy.vix@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
apache-airflow = "^2.9.1"
ruff = "^0.4.6"
pandas = "^2.2.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
